Manual parent dropdown
To help improve loadtime of children devices in environments with alot of chassis.  The Parent Device dropdown won't be populated until the first time the dropdown is clicked.  In the future, this is probably another place where the Select HTML can be stored in the fac_DataCache whenever a chassis is added/removed/updated.